+++ src/falconpy/_auth_object/_falcon_auth.py
@@ -39,72 +39,11 @@ OTHER DEALINGS IN THE SOFTWARE.
 For more information, please refer to <https://unlicense.org>
 """
 import time
-from abc import ABC, abstractmethod
 from typing import Dict, Optional
-from ._token_fail_reason import TokenFailReason
-from ._util import generate_b64cred, autodiscover_region, perform_request, generate_error_result
-from ._endpoint._oauth2 import _oauth2_endpoints as AuthEndpoints
-
-
-class BaseFalconAuth(ABC):
-    """Abstract class to provide an interface to the CrowdStrike Falcon OAuth2 API.
-
-    This class does not implement a generic constructor and is not intended to be used by
-    developers directly. You must work with a derivative of this class, such as a FalconAuth object.
-    """
-    #  ______  _______ _______ _______ _     _        _______
-    #  |     \ |______ |______ |_____| |     | |         |
-    #  |_____/ |______ |       |     | |_____| |_____    |
-
-    #  _______ _______ _______ _     _  _____  ______  _______
-    #  |  |  | |______    |    |_____| |     | |     \ |______
-    #  |  |  | |______    |    |     | |_____| |_____/ ______|
-
-    # The generic login and logout handlers must be individually defined by all
-    # inheriting classes. The private methods defined here are used to allow for
-    # easy overridding of login and logout processing by inheriting classes without
-    # altering the parent handler method that may be leveraged by other inheriting
-    # class types.
-    @abstractmethod
-    def _login_handler(self) -> dict or bool:
-        """Login to the Falcon API by requesting a new token."""
-
-    @abstractmethod
-    def _logout_handler(self) -> dict or bool:
-        """Log out of the Falcon API by revoking the current token."""
-
-    @abstractmethod
-    def login(self) -> dict or bool:
-        """Generic login handler interface."""
-
-    @abstractmethod
-    def logout(self) -> dict or bool:
-        """Generic logout handler interface."""
-
-    #   _____   ______  _____   _____  _______  ______ _______ _____ _______ _______
-    #  |_____] |_____/ |     | |_____] |______ |_____/    |      |   |______ |______
-    #  |       |    \_ |_____| |       |______ |    \_    |    __|__ |______ ______|
-    #
-    # These properties are present within all BaseFalconAuth derivatives.
-    @property
-    @abstractmethod
-    def auth_headers(self) -> Dict[str, str]:
-        """Get a dictionary of headers that can authenticate an HTTP request."""
-
-    @property
-    @abstractmethod
-    def authenticated(self) -> bool:
-        """Read-only property to return whether authentication is successful."""
-
-    @property
-    @abstractmethod
-    def token_expired(self) -> bool:
-        """Read-only property that returns the current token expiration status."""
-
-    @property
-    @abstractmethod
-    def cred_format_valid(self) -> bool:
-        """Read-only property that returns a boolean if the creds dictionary is valid."""
+from ._base_falcon_auth import BaseFalconAuth
+from .._enum import TokenFailReason
+from .._util import generate_b64cred, autodiscover_region, perform_request, generate_error_result
+from .._endpoint._oauth2 import _oauth2_endpoints as AuthEndpoints
 
 
 class FalconAuth(BaseFalconAuth):
@@ -291,120 +230,3 @@ class FalconAuth(BaseFalconAuth):
             self.login()
 
         return {"Authorization": f"Bearer {self.token_value}"}
-
-
-class UberInterface(FalconAuth):
-    """Uber Class specific interface."""
-
-    # ____ ___ ___ ____ _ ___  _  _ ___ ____ ____
-    # |__|  |   |  |__/ | |__] |  |  |  |___ [__
-    # |  |  |   |  |  \ | |__] |__|  |  |___ ___]
-    #
-    # Attributes present only within the Uber Class.
-    #
-    # A dictionary of every available API operation provided by the library.
-    commands: dict = {}
-
-    # ____ ____ _  _ ____ ___ ____ _  _ ____ ___ ____ ____
-    # |    |  | |\ | [__   |  |__/ |  | |     |  |  | |__/
-    # |___ |__| | \| ___]  |  |  \ |__| |___  |  |__| |  \
-    #
-    # Starting in v1.3.0, the Uber Class constructs itself leveraging the generic
-    # FalconAuth constructor. This results in the Uber Class benefiting from a new
-    # authentication style; Legacy / Token authentication.
-    def __init__(self,
-                 access_token: Optional[str or bool] = False,
-                 base_url: Optional[str] = "https://api.crowdstrike.com",
-                 creds: Optional[dict] = None,
-                 client_id: Optional[str] = None,
-                 client_secret: Optional[str] = None,
-                 member_cid: Optional[str] = None,
-                 ssl_verify: Optional[bool] = True,
-                 proxy: Optional[dict] = None,
-                 timeout: Optional[float or tuple] = None,
-                 user_agent: Optional[str] = None,
-                 renew_window: Optional[int] = 120
-                 ) -> "UberInterface":
-        """Construct an instance of the UberInterface class."""
-        super().__init__(base_url=base_url,
-                         ssl_verify=ssl_verify,
-                         timeout=timeout,
-                         proxy=proxy,
-                         user_agent=user_agent,
-                         access_token=access_token,
-                         creds=creds,
-                         client_id=client_id,
-                         client_secret=client_secret,
-                         member_cid=member_cid,
-                         renew_window=renew_window
-                         )
-
-    # _  _ ____ ___ _  _ ____ ___  ____
-    # |\/| |___  |  |__| |  | |  \ [__
-    # |  | |___  |  |  | |__| |__/ ___]
-    #
-    # Override the default login and logout handlers to
-    # provide Uber Class-specific functionality.
-    def _login_handler(self) -> bool:
-        """Generate an authorization token."""
-        super()._login_handler()
-
-        return self.authenticated
-
-    def _logout_handler(self) -> bool:
-        """Revoke the current authorization token."""
-        result = super()._logout_handler()
-
-        return bool(result["status_code"] == 200)
-
-    # _    ____ ____ ____ ____ _   _    _  _ ____ _  _ ___  _    ____ ____ ____
-    # |    |___ | __ |__| |     \_/     |__| |__| |\ | |  \ |    |___ |__/ [__
-    # |___ |___ |__] |  | |___   |      |  | |  | | \| |__/ |___ |___ |  \ ___]
-    #
-    # These handlers provide legacy Uber Class-specific functionality that will be
-    # maintained for provide backwards compatibility purposes.
-    def authenticate(self) -> bool:
-        """Legacy Uber Class functionality handler.
-
-        DEPRECATED
-        ----
-        Consider updating your code to leverage the login method.
-        """
-        return super().login()
-
-    def deauthenticate(self) -> bool:
-        """Legacy Uber Class functionality handler.
-
-        DEPRECATED
-        ----
-        Consider updating your code to leverage the logout method.
-        """
-        return super().logout()
-
-    def valid_cred_format(self) -> bool:
-        """Legacy property to confirm credential dictionary format.
-        
-        DEPRECATED
-        ----
-        Consider updating your code to leverage the cred_format_valid property.
-        """
-        return self.cred_format_valid
-
-    def headers(self) -> Dict[str, str]:
-        """Legacy property getter for the current authorization headers.
-
-        DEPRECATED
-        ----
-        Consider updating your code to leverage the auth_headers property.
-        """
-        return self.auth_headers
-
-    @property
-    def token(self) -> str:
-        """Legacy attribute handler to return the token string.
-
-        DEPRECATED
-        ----
-        Consider updating your code to leverage the token_value property.
-        """
-        return self.token_value
