+++ src/falconpy/_result/__init__.py
@@ -1,0 +1,4 @@
+from ._result import Result
+from ._expanded_result import ExpandedResult
+
+__all__ = ["Result", "ExpandedResult"]
