+++ src/falconpy/_service_class/__init__.py
@@ -1,0 +1,4 @@
+from ._base_service_class import BaseServiceClass
+from ._service_class import ServiceClass
+
+__all__ = ["BaseServiceClass", "ServiceClass"]
