+++ src/falconpy/_util/__init__.py
@@ -1,0 +1,36 @@
+from ._functions import (
+    validate_payload,
+    generate_b64cred,
+    handle_single_argument,
+    force_default,
+    service_request,
+    perform_request,
+    generate_error_result,
+    generate_ok_result,
+    get_default,
+    args_to_params,
+    process_service_request,
+    confirm_base_url,
+    confirm_base_region,
+    return_preferred_default,
+    base_url_regions,
+    autodiscover_region,
+    _ALLOWED_METHODS
+)
+
+from ._uber import (
+    create_uber_header_payload,
+    handle_body_payload_ids,
+    scrub_target,
+    handle_container_operations,
+    uber_request_keywords
+)
+
+__all__ = ["create_uber_header_payload", "handle_body_payload_ids", "scrub_target",
+           "handle_container_operations", "uber_request_keywords", "autodiscover_region",
+           "validate_payload", "generate_b64cred", "handle_single_argument", "force_default",
+           "service_request", "perform_request", "generate_error_result", "generate_ok_result",
+           "get_default", "args_to_params", "process_service_request", "confirm_base_url",
+           "confirm_base_region", "return_preferred_default", "base_url_regions",
+           "_ALLOWED_METHODS"
+           ]
